Add NewFromRat helper function (#6)
This pull request drops the fork that we have created in favour of an utility function that does the same thing. The reasoning goes like this: it does not seem that the maintainer is interested and, furthermore, we wouldn't need to maintain a fork.

In general, we should refactor code in such a way that would not need Rat. Afterwards, we will no longer need this function.